requirements.txt
streamlit>=1.36.0
requests>=2.31.0
python-dotenv>=1.0.1
flask>=3.0.0
gunicorn>=22.0.0
